pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "customtkinter>=5.2.2",
    "flask-login>=0.6.3",
    "flask-wtf>=1.2.2",
    "mss>=10.0.0",
    "numpy>=2.2.3",
    "oauthlib>=3.2.2",
    "openai>=1.63.2",
    "opencv-python>=4.11.0.86",
    "pillow>=11.1.0",
    "pyautogui>=0.9.54",
    "python-dotenv>=1.0.1",
    "torch>=2.6.0",
    "torchvision>=0.21.0",
    "twilio>=9.4.6",
]

[[tool.uv.index]]
explicit = true
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"

[tool.uv.sources]
torch = [{ index = "pytorch-cpu", marker = "platform_system == 'Linux'" }]
torchvision = [{ index = "pytorch-cpu", marker = "platform_system == 'Linux'" }]
